requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
openai>=1

# Optional: spelling correction and testing
pyspellchecker
pytest
httpx

python-dotenv
